+++ src/test/java/com/syncleus/ferma/FramedVertexTest.java
@@ -32,8 +32,7 @@ import com.google.common.collect.Lists;
 import javax.annotation.Nullable;
 
 public class FramedVertexTest {
-    private static final Function<GraphTraversal<? extends Vertex, ? extends Vertex>, GraphTraversal<?, ?>>
-        OUT_TRAVERSAL = input -> input.out();
+    private static final Function<GraphTraversal<Vertex, Vertex>, GraphTraversal<?, ?>> OUT_TRAVERSAL = input -> input.out();
 
     private FramedGraph fg;
     private Person p1;